requirements.txt
ntplib==0.4.0
Flask==2.3.3
SQLAlchemy==2.0.20
requests==2.31.0
PyYAML==6.0.1
python-dateutil==2.8.2
psutil==5.9.5

